python/monitor/requirements.txt
fastapi
uvicorn
aioping
httpx
speedtest-cli
SQLAlchemy
pydantic
python-dotenv
